Pushing this so I don't lose it
This breaks the tests. I know. But it also contains *features* that
should not be lost. And it's Worked Wonderfully For Me™.

It now is able to retrieve assets mentioned inside CSS files, and keeps
a local cache based on URL to better deal with flaky networks. Downside:
if the resource at the URL changes you need to delete the cache.

Will clean this up and push a proper release of Slippery. Bug me if it's
end of 2014 or later and I haven't yet.